requirements.txt
amqp==5.3.1
asgiref==3.9.2
async-timeout==5.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2025.8.3
chardet==4.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.2.18
Django==4.1.2
gunicorn==20.1.0
idna==2.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2025.2
redis==4.3.4
requests==2.25.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.14
wrapt==1.17.3
whitenoise==6.11.0
pytest==8.3.2
pytest-django==4.9.0
pytest-cov==5.0.0
